requirements.txt
# Core ML libraries
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0

# PDF processing
PyPDF2>=3.0.0

# Data processing
numpy>=1.24.0

# Text and chunking
textblob>=0.17.1
langchain-text-splitters>=0.0.1
rank-bm25>=0.2.2

# Caching
diskcache>=5.6.3

# Utilities
pathlib2>=2.3.0
tqdm>=4.65.0
psutil>=5.9.0
